initial refactor and migrate to typescript
