+++ ci/codefresh.yaml
@@ -1,4 +1,4 @@
-version: "1.0"
+version: '1.0'
 
 stages:
   - Clone
