+++ deno.lock
@@ -1,0 +1,80 @@
+{
+  "version": "4",
+  "specifiers": {
+    "jsr:@cloudydeno/kubernetes-apis@~0.5.1": "0.5.2",
+    "jsr:@cloudydeno/kubernetes-client@0.7.3": "0.7.3",
+    "jsr:@cloudydeno/kubernetes-client@~0.7.3": "0.7.3",
+    "jsr:@fakoua/zip-ts@^1.3.1": "1.3.1",
+    "jsr:@std/assert@~0.218.2": "0.218.2",
+    "jsr:@std/bytes@^1.0.2": "1.0.2",
+    "jsr:@std/fs@0.218.2": "0.218.2",
+    "jsr:@std/path@0.218.2": "0.218.2",
+    "jsr:@std/path@1.0.6": "1.0.6",
+    "jsr:@std/path@~0.218.2": "0.218.2",
+    "jsr:@std/streams@1.0.5": "1.0.5",
+    "jsr:@std/yaml@1.0.5": "1.0.5",
+    "jsr:@std/yaml@^1.0.5": "1.0.5"
+  },
+  "jsr": {
+    "@cloudydeno/kubernetes-apis@0.5.2": {
+      "integrity": "f5fd9c94987f54fde72c6103fa0663ccb71cb6e328e7be44af1d59145adb31ee",
+      "dependencies": [
+        "jsr:@cloudydeno/kubernetes-client@0.7.3"
+      ]
+    },
+    "@cloudydeno/kubernetes-client@0.7.3": {
+      "integrity": "eb9fd12ac6c58e7995cc2d41a72ea9d153970900e184a5bd6fa2f2b19da5ec1a",
+      "dependencies": [
+        "jsr:@std/path@1.0.6",
+        "jsr:@std/streams",
+        "jsr:@std/yaml@1.0.5"
+      ]
+    },
+    "@fakoua/zip-ts@1.3.1": {
+      "integrity": "b66f813015a59f40143c2bac337aed9abac7827da5e39c34549108aa129f33e7",
+      "dependencies": [
+        "jsr:@std/fs",
+        "jsr:@std/path@0.218.2"
+      ]
+    },
+    "@std/assert@0.218.2": {
+      "integrity": "7f0a5a1a8cf86607cd6c2c030584096e1ffad27fc9271429a8cb48cfbdee5eaf"
+    },
+    "@std/bytes@1.0.2": {
+      "integrity": "fbdee322bbd8c599a6af186a1603b3355e59a5fb1baa139f8f4c3c9a1b3e3d57"
+    },
+    "@std/fs@0.218.2": {
+      "integrity": "dd9431453f7282e8c577cc22c9e6d036055a9a980b5549f887d6012969fabcca",
+      "dependencies": [
+        "jsr:@std/assert",
+        "jsr:@std/path@~0.218.2"
+      ]
+    },
+    "@std/path@0.218.2": {
+      "integrity": "b568fd923d9e53ad76d17c513e7310bda8e755a3e825e6289a0ce536404e2662",
+      "dependencies": [
+        "jsr:@std/assert"
+      ]
+    },
+    "@std/path@1.0.6": {
+      "integrity": "ab2c55f902b380cf28e0eec501b4906e4c1960d13f00e11cfbcd21de15f18fed"
+    },
+    "@std/streams@1.0.5": {
+      "integrity": "74e5c73d7d68eeab0d7fba3b05cbee3ba4ac5ae37c5f4e675f67e62f8f53edc4",
+      "dependencies": [
+        "jsr:@std/bytes"
+      ]
+    },
+    "@std/yaml@1.0.5": {
+      "integrity": "71ba3d334305ee2149391931508b2c293a8490f94a337eef3a09cade1a2a2742"
+    }
+  },
+  "workspace": {
+    "dependencies": [
+      "jsr:@cloudydeno/kubernetes-apis@~0.5.1",
+      "jsr:@cloudydeno/kubernetes-client@~0.7.3",
+      "jsr:@fakoua/zip-ts@^1.3.1",
+      "jsr:@std/yaml@^1.0.5"
+    ]
+  }
+}
