+++ src/utils/runtime-type.ts
@@ -1,0 +1,5 @@
+export enum RuntimeType {
+  pipelines = 'Pipelines Runtime',
+  gitops = 'GitOps Runtime',
+  onprem = 'On-Prem'
+}
